+++ src/syntax/c/c_parser.py
@@ -103,8 +103,8 @@ class CParser(Parser):
             self._tokens.append(token)
 
         parser_state = CParserState()
-        parser_state.continuation_state = 1 if lexer_state.in_block_comment else 0
-        parser_state.parsing_continuation = lexer_state.in_block_comment
+        parser_state.parsing_continuation = lexer_state.in_block_comment or lexer_state.in_preprocessor
+        parser_state.continuation_state = 1 if lexer_state.in_block_comment else (2 if lexer_state.in_preprocessor else 0)
         parser_state.lexer_state = lexer_state
         parser_state.in_element = in_element
         return parser_state